pyproject.toml
[project]
name = "address-geocoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.9.0",
    "chardet>=5.2.0",
    "click>=8.1.8",
    "ipython>=8.18.1",
    "passyunk",
    "polars>=1.34.0",
    "pyarrow>=14",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "retrying>=1.4.2",
    "ruff>=0.14.9",
    "setuptools>=80.9.0",
    "usaddress>=0.5.16",
    "wheel>=0.46.2",
]

[tool.uv.sources]
passyunk = { git = "https://github.com/CityOfPhiladelphia/passyunk" }
